requirements.txt
customtkinter
cryptography
wmi
argon2-cffi
pywin32
tkinter